Rename resource_manager to remote_host
